requirements.txt
streamlit>=1.28.0
whisper-openai>=20231117
pyaudio>=0.2.11
numpy>=1.21.0
scipy>=1.7.0
librosa>=0.9.0
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
SpeechRecognition>=3.10.0
whisper>=1.1.10 